@@ -2,9 +2,15 @@
 /* global ajaxurl */
 /* global require */
 
+import "./helpers/babel-polyfill";
+
 var Jed = require( "jed" );
-var Paper = require( "yoastseo/js/values/Paper" );
-var SEOAssessor = require( "yoastseo/js/seoAssessor" );
+
+import {
+	Paper,
+	SEOAssessor,
+} from "yoastseo";
+
 var TaxonomyAssessor = require( "./assessors/taxonomyAssessor" );
 var isUndefined = require( "lodash/isUndefined" );
 
