@@ -3,8 +3,20 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
+import styled from "styled-components";
+import { __ } from "@wordpress/i18n";
 
 import Results from "./Results";
+import Collapsible from "../SidebarCollapsible";
+import getIndicatorForScore from "../../analysis/getIndicatorForScore";
+import { getIconForScore } from "./mapResults";
+
+const AnalysisHeader = styled.span`
+	font-size: 1em;
+	font-weight: bold;
+	margin: 0 0 8px;
+	display: block;
+`;
 
 let localizedData = {};
 if( window.wpseoPostScraperL10n ) {
@@ -18,16 +30,27 @@ if( window.wpseoPostScraperL10n ) {
  */
 class ReadabilityAnalysis extends React.Component {
 	render() {
+		const score = getIndicatorForScore( this.props.overallScore );
 		return (
-			<Results
-				canChangeLanguage={ ! ( localizedData.settings_link === "" ) }
-				showLanguageNotice={ true }
-				changeLanguageLink={ localizedData.settings_link }
-				language={ localizedData.language }
-				results={ this.props.results }
-				marksButtonClassName="yoast-tooltip yoast-tooltip-s"
-				marksButtonStatus={ this.props.marksButtonStatus }
-			/>
+			<Collapsible
+				title={ __( "Readability", "wordpress-seo" ) }
+				titleScreenReaderText={ score.screenReaderReadabilityText }
+				prefixIcon={ getIconForScore( score.className ) }
+				prefixIconCollapsed={ getIconForScore( score.className ) }
+			>
+				<AnalysisHeader>
+					Analysis results:
+				</AnalysisHeader>
+				<Results
+					canChangeLanguage={ ! ( localizedData.settings_link === "" ) }
+					showLanguageNotice={ true }
+					changeLanguageLink={ localizedData.settings_link }
+					language={ localizedData.language }
+					results={ this.props.results }
+					marksButtonClassName="yoast-tooltip yoast-tooltip-s"
+					marksButtonStatus={ this.props.marksButtonStatus }
+				/>
+			</Collapsible>
 		);
 	}
 }
@@ -36,6 +59,7 @@ ReadabilityAnalysis.propTypes = {
 	results: PropTypes.array,
 	marksButtonStatus: PropTypes.string,
 	hideMarksButtons: PropTypes.bool,
+	overallScore: PropTypes.number,
 };
 
 /**
@@ -52,6 +76,7 @@ function mapStateToProps( state, ownProps ) {
 	return {
 		results: state.analysis.readability.results,
 		marksButtonStatus: marksButtonStatus,
+		overallScore: state.analysis.readability.overallScore,
 	};
 }
 
