@@ -4,17 +4,20 @@ import get from "lodash/get";
 /**
  * Sets the l10n for the given textdomain in Jed.
  *
- * All the locale data should be available in wpseoYoastComponentsL10n
+ * All the locale data should be available in the l10nNamespace.
  *
- * @param {string} textdomain The textdomain to set the locale data for.
+ * @param {string} textdomain      The textdomain to set the locale data for.
+ * @param {string} [l10nNamespace] The global namespace to get the localization
+ *                                 data from.
+ *
  * @returns {void}
  */
-export function setTextdomainL10n( textdomain ) {
+export function setTextdomainL10n( textdomain, l10nNamespace = "wpseoYoastJSL10n" ) {
 	const jed = getI18n();
 	const currentTranslations = get( jed, [ "options", "locale_data", textdomain ], false );
 
 	if ( currentTranslations === false ) {
-		const translations = get( window, [ "wpseoYoastJSL10n", textdomain, "locale_data", textdomain ], false );
+		const translations = get( window, [ l10nNamespace, textdomain, "locale_data", textdomain ], false );
 
 		if ( translations === false ) {
 			// Jed needs to have meta information in the object keyed by an empty string.
