@@ -1,4 +1,4 @@
-import * as activeKeyword from "./activeKeyword";
+import * as focusKeyword from "./focusKeyword";
 import * as activeTab from "./activeTab";
 import * as cornerstoneContent from "./cornerstoneContent";
 import * as keywords from "./keywords";
@@ -6,7 +6,7 @@ import * as markerButtons from "./markerButtons";
 import * as openSidebarSections from "./openSidebarSections";
 
 export default {
-	activeKeyword,
+	focusKeyword,
 	activeTab,
 	cornerstoneContent,
 	keywords,
