@@ -76,7 +76,7 @@ var imageScoreClass = "image yoast-logo svg";
 	 *
 	 * @returns {void}
 	 */
-	function initialise() {
+	function initialize() {
 		var notAvailableStatus = "na";
 
 		if ( wpseoPostScraperL10n.contentAnalysisActive === "1" ) {
@@ -89,7 +89,7 @@ var imageScoreClass = "image yoast-logo svg";
 	}
 
 	module.exports = {
-		initalise: initialise,
+		initialize: initialize,
 		updateScore: updateScoreInPublishBox,
 	};
 }( jQuery ) );
