@@ -1,4 +1,4 @@
-import { isGutenbergDataAvailable, isGutenbergPostAvailable } from "../src/helpers/isGutenbergAvailable.js";
+import { isGutenbergDataAvailable } from "../src/helpers/isGutenbergAvailable.js";
 
 describe( "isGutenbergDataAvailable", () => {
 	it( "returns true if both wp and wp.data are defined", () => {
@@ -25,18 +25,5 @@ describe( "isGutenbergDataAvailable", () => {
 		window.wp = undefined;
 		const actual = isGutenbergDataAvailable();
 		expect( actual ).toBe( false );
-	} );
-} );
-
-describe( 'isGutenbergPostAvailable', () => {
-	it( 'returns true if _wpGutenbergPost is defined', () => {
-		window._wpGutenbergPost = { id: 1234 };
-		const actual = isGutenbergPostAvailable();
-		expect( actual ).toBe( true );
-	} );
-	it( 'returns false if _wpGutenbergPost is not defined', () => {
-		delete window._wpGutenbergPost;
-		const actual = isGutenbergPostAvailable();
-		expect( actual ).toEqual( false );
 	} );
 } );
