@@ -57,7 +57,7 @@ final class Test_Admin_Asset_Dev_Server_Location extends PHPUnit_Framework_TestC
 	public function test_get_url_default() {
 		$asset = new WPSEO_Admin_Asset( array(
 			'name' => 'commons',
-			'src'  => 'commons',
+			'src'  => 'select2',
 		) );
 
 		$dev_server_location = new WPSEO_Admin_Asset_Dev_Server_Location();
@@ -65,29 +65,6 @@ final class Test_Admin_Asset_Dev_Server_Location extends PHPUnit_Framework_TestC
 
 		$actual   = $dev_server_location->get_url( $asset, WPSEO_Admin_Asset::TYPE_CSS );
 		$expected = $default_location->get_url( $asset, WPSEO_Admin_Asset::TYPE_CSS );
-
-		$this->assertEquals( $expected, $actual );
-	}
-
-	/**
-	 * Tests that the dev server falls back to the default asset if it isn't on the dev server.
-	 *
-	 * @integration_test
-	 *
-	 * @covers WPSEO_Admin_Asset_Dev_Server_Location::get_url()
-	 * @covers WPSEO_Admin_Asset_SEO_Location::get_url()
-	 */
-	public function test_get_url_default_2() {
-		$asset = new WPSEO_Admin_Asset( array(
-			'name' => 'other',
-			'src'  => 'other',
-		) );
-
-		$dev_server_location = new WPSEO_Admin_Asset_Dev_Server_Location();
-		$default_location    = new WPSEO_Admin_Asset_SEO_Location( WPSEO_FILE );
-
-		$actual   = $dev_server_location->get_url( $asset, WPSEO_Admin_Asset::TYPE_JS );
-		$expected = $default_location->get_url( $asset, WPSEO_Admin_Asset::TYPE_JS );
 
 		$this->assertEquals( $expected, $actual );
 	}
