@@ -15,7 +15,7 @@ if ( ! function_exists( 'add_filter' ) ) {
  * {@internal Nobody should be able to overrule the real version number as this can cause
  *            serious issues with the options, so no if ( ! defined() ).}}
  */
-define( 'WPSEO_VERSION', '7.9.1' );
+define( 'WPSEO_VERSION', '8.0-RC3' );
 
 
 if ( ! defined( 'WPSEO_PATH' ) ) {
